split: set, schedule, exercise
