+++ apps/web/src/entities/exercise/api/useGetExerciseFinishHistory.ts
@@ -1,4 +1,3 @@
-import type { MockedResponse } from '@apollo/client/testing';
 import { useSuspenseQuery } from '@apollo/client'
 import GetExerciseFinishHistory from '@entities/exercise/api/graphql/query/GetExerciseFinishHistory';
 
@@ -9,20 +8,4 @@ export default function useGetExerciseFinishHistory(exerciseId: number) {
   >(GetExerciseFinishHistory, {
     variables: { exerciseId: Number(exerciseId) }
   })
-}
-
-export const GetExerciseFinishHistoryMock: MockedResponse<
-  GetExerciseFinishHistoryResponse,
-  GetExerciseFinishHistoryVariable
-> = {
-  request: {
-    query: GetExerciseFinishHistory,
-  },
-  result: () => {
-    return {
-      data: {
-        getExerciseFinishHistory: []
-      }
-    }
-  }
 }
