+++ apps/web/src/entities/exercise/api/useGetExerciseListByExercisePresetId.ts
@@ -1,6 +1,4 @@
-import type { MockedResponse } from '@apollo/client/testing';
 import { useSuspenseQuery } from '@apollo/client'
-import { FitnessMockData } from '../../fitness/api';
 import GetExerciseListByExercisePresetId from './graphql/query/GetExerciseListByExercisePresetId';
 
 export default function useGetExerciseListByExercisePresetId(exercisePresetId: number) {
@@ -10,35 +8,4 @@ export default function useGetExerciseListByExercisePresetId(exercisePresetId: n
   >(GetExerciseListByExercisePresetId, {
     variables: { exercisePresetId: Number(exercisePresetId) }
   })
-}
-
-export const GetExerciseListByExercisePresetIdMock: MockedResponse<
-  GetExerciseByExercisePresetIdResponse,
-  GetExerciseByExercisePresetIdVariable
-> = {
-  request: {
-    query: GetExerciseListByExercisePresetId,
-  },
-  result: () => {
-    return {
-      data: {
-        getExerciseListByExercisePresetId: [
-          {
-            id: 1,
-            exercise: 1,
-            deps: 0,
-            __typename: 'ExerciseWithFitness',
-            fitness: FitnessMockData[0]
-          },
-          {
-            id: 2,
-            exercise: 2,
-            deps: 0,
-            __typename: 'ExerciseWithFitness',
-            fitness: FitnessMockData[1]
-          }
-        ]
-      }
-    }
-  }
 }
