+++ apps/web/src/entities/exercise/api/useGetExerciseListByScheduleId.ts
@@ -1,6 +1,4 @@
-import type { MockedResponse } from '@apollo/client/testing';
 import { useSuspenseQuery } from '@apollo/client'
-import { FitnessMockData } from '@entities/fitness/api';
 import GetExerciseListByScheduleId from '../api/graphql/query/GetExerciseListByScheduleId';
 
 export default function useGetExerciseListByScheduleId(scheduleId: number) {
@@ -10,35 +8,4 @@ export default function useGetExerciseListByScheduleId(scheduleId: number) {
   >(GetExerciseListByScheduleId, {
     variables: { scheduleId }
   })
-}
-
-export const GetExerciseListByScheduleIdMock: MockedResponse<
-  GetExerciseListByScheduleIdResponse,
-  GetExerciseListByScheduleIdVariable
-> = {
-  request: {
-    query: GetExerciseListByScheduleId,
-  },
-  result: () => {
-    return {
-      data: {
-        getExerciseListByScheduleId: [
-          {
-            id: 1,
-            exercise: 1,
-            deps: 0,
-            __typename: 'ExerciseWithFitness',
-            fitness: FitnessMockData[0]
-          },
-          {
-            id: 2,
-            exercise: 2,
-            deps: 0,
-            __typename: 'ExerciseWithFitness',
-            fitness: FitnessMockData[1]
-          }
-        ]
-      }
-    }
-  }
 }
