+++ apps/web/src/entities/schedule/api/useCloneSchedule.mock.ts
@@ -1,0 +1,25 @@
+import type { MockedResponse } from '@apollo/client/testing';
+import type { Schedule } from 'fitness-struct'
+import CloneSchedule from './graphql/mutation/CloneScheduleGql';
+import { ScheduleMockData } from '.'
+
+export const CloneScheduleMock: MockedResponse<CloneScheduleResponse, CloneScheduleVariable> = {
+  request: {
+    query: CloneSchedule
+  },
+  result: (v) => {
+    const id = Math.max(...Object.keys(ScheduleMockData).map(Number)) + 1
+    ScheduleMockData[id] = {
+      ...(ScheduleMockData[v.id] as Schedule.Schedule),
+      id,
+      year: v.targetDate.year,
+      month: v.targetDate.month,
+      date: v.targetDate.date
+    }
+    return {
+      data: {
+        cloneSchedule: ScheduleMockData[id]
+      }
+    }
+  }
+}
