+++ apps/web/src/entities/schedule/api/useCloneScheduleFromPreset.ts
@@ -1,8 +1,4 @@
-import type { MockedResponse } from '@apollo/client/testing';
-import type { ExercisePreset } from 'fitness-struct'
 import { useMutation } from '@apollo/client'
-import { ExercisePresetMockData } from '@entities/exercisePreset/api';
-import { ScheduleMockData } from '@entities/schedule/api';
 import CloneScheduleFromPreset from '@entities/schedule/api/graphql/mutation/CloneScheduleFromPresetGql';
 
 export default function useCloneScheduleFromPreset() {
@@ -21,29 +17,4 @@ export default function useCloneScheduleFromPreset() {
       })
     }
   })
-}
-export const CloneScheduleFromPresetMock: MockedResponse<CloneScheduleFromPresetResponse, CloneScheduleFromPresetVariable> = {
-  request: {
-    query: CloneScheduleFromPreset
-  },
-  result: (v) => {
-    const id = Math.max(...Object.keys(ScheduleMockData).map(Number)) + 1
-    ScheduleMockData[id] = {
-      ...(ExercisePresetMockData[v.presetId] as ExercisePreset.Preset),
-      beforeTime: 0,
-      start: 0,
-      breakTime: 0,
-      workoutTimes: 0,
-      type: 'SCHEDULED',
-      id,
-      year: v.targetDate.year,
-      month: v.targetDate.month,
-      date: v.targetDate.date
-    }
-    return {
-      data: {
-        cloneScheduleFromPreset: ScheduleMockData[id]
-      }
-    }
-  }
 }
