+++ apps/web/src/entities/schedule/api/useCreateSchedule.ts
@@ -1,6 +1,4 @@
-import type { MockedResponse } from '@apollo/client/testing';
 import { useMutation } from '@apollo/client'
-import { ScheduleMockData } from '@entities/schedule/api';
 import CreateSchedule from '@entities/schedule/api/graphql/mutation/CreateScheduleGql';
 
 export default function useCreateSchedule() {
@@ -29,21 +27,4 @@ export default function useCreateSchedule() {
       })
     }
   })
-}
-export const CreateScheduleMock: MockedResponse<CreateScheduleResponse, CreaetScheduleVariable> = {
-  request: {
-    query: CreateSchedule
-  },
-  result: (v) => {
-    const id = Math.max(...Object.keys(ScheduleMockData).map(Number)) + 1
-    ScheduleMockData[id] = {
-      ...v.createSchedule,
-      id: id
-    }
-    return {
-      data: {
-        createSchedule: ScheduleMockData[id]
-      }
-    }
-  }
 }
