+++ apps/web/src/entities/schedule/api/useDeleteSchedule.mock.ts
@@ -1,0 +1,17 @@
+import type { MockedResponse } from '@apollo/client/testing';
+import { ScheduleMockData } from '@entities/schedule/api';
+import DeleteScheduleGql from '@entities/schedule/api/graphql/mutation/DeleteScheduleGql';
+
+export const DeleteScheduleMock: MockedResponse<DeleteScheduleResponse, DeleteScheduleVariable> = {
+  request: {
+    query: DeleteScheduleGql
+  },
+  result: (v) => {
+    delete ScheduleMockData[v.id]
+    return {
+      data: {
+        deleteSchedule: `delete - schedule - ${v.id}`
+      }
+    }
+  }
+}
