+++ apps/web/src/entities/schedule/api/useDeleteSchedule.ts
@@ -1,6 +1,4 @@
-import type { MockedResponse } from '@apollo/client/testing';
 import { useMutation } from '@apollo/client'
-import { ScheduleMockData } from '@entities/schedule/api';
 import DeleteScheduleGql from '@entities/schedule/api/graphql/mutation/DeleteScheduleGql';
 
 export default function useDeleteSchedule() {
@@ -21,17 +19,4 @@ export default function useDeleteSchedule() {
       })
     }
   })
-}
-export const DeleteScheduleMock: MockedResponse<DeleteScheduleResponse, DeleteScheduleVariable> = {
-  request: {
-    query: DeleteScheduleGql
-  },
-  result: (v) => {
-    delete ScheduleMockData[v.id]
-    return {
-      data: {
-        deleteSchedule: `delete - schedule - ${v.id}`
-      }
-    }
-  }
 }
