+++ apps/web/src/entities/schedule/api/useGetScheduleByDate.mock.ts
@@ -1,0 +1,16 @@
+import type { MockedResponse } from '@apollo/client/testing';
+import GetScheduleByDateGql from './graphql/query/GetScheduleByDateGql';
+import { ScheduleMockData } from '.'
+
+export const GetScheduleByDateMock: MockedResponse<GetScheduleByDateResponse, GetScheduleByDateVariable> = {
+  request: {
+    query: GetScheduleByDateGql
+  },
+  result: (v) => {
+    return {
+      data: {
+        getScheduleByDate: Object.values(ScheduleMockData).filter(o => o.year === v.year && o.month === v.month && o.date === v.date)
+      }
+    }
+  }
+}
