+++ apps/web/src/entities/schedule/api/useGetScheduleById.mock.ts
@@ -1,0 +1,16 @@
+import type { MockedResponse } from '@apollo/client/testing';
+import GetScheduleByIdGql from '../api/graphql/query/GetScheduleByIdGql';
+import { ScheduleMockData } from '.';
+
+export const GetScheduleByIdMock: MockedResponse<GetScheduleByIdResponse, GetScheduleByIdVAriable> = {
+  request: {
+    query: GetScheduleByIdGql
+  },
+  result: (v) => {
+    return {
+      data: {
+        getScheduleById: ScheduleMockData[v.id]
+      }
+    }
+  }
+}
