+++ apps/web/src/entities/schedule/api/useGetScheduleStatusByMonth.ts
@@ -1,7 +1,4 @@
-import type { MockedResponse } from '@apollo/client/testing';
-import type { Schedule } from 'fitness-struct';
 import { useSuspenseQuery } from '@apollo/client'
-import { ScheduleMockData } from '@entities/schedule/api';
 import GetScheduleStatusByMonthGql from '@entities/schedule/api/graphql/query/GetScheduleStatusByMonthGql';
 
 export default function useGetScheduleStatusByMonth(year: number, month: number) {
@@ -12,26 +9,4 @@ export default function useGetScheduleStatusByMonth(year: number, month: number)
         month
       }
     })
-}
-
-export const GetScheduleStateByDateMock: MockedResponse<
-  GetScheduleStatusByMonthResponse,
-  GetScheduleStatusByMonthVariable
-> = {
-  request: {
-    query: GetScheduleStatusByMonthGql
-  },
-  result: (v) => {
-    const list = Object.values(ScheduleMockData).filter(o => o.year === v.year && o.month === v.month)
-    list.sort((a, b) => a.date - b.date)
-    return {
-      data: {
-        getScheduleStatusByMonth: list.reduce((acc, cur) => {
-          const temp = acc[cur.date] ?? []
-          temp.push(cur.type)
-          return acc
-        }, [] as Schedule.IType[][])
-      }
-    }
-  }
 }
