+++ apps/web/src/entities/set/api/useCreateSet.mock.ts
@@ -1,0 +1,21 @@
+import type { MockedResponse } from '@apollo/client/testing';
+import { SetMockData } from '@entities/set/api';
+import CreateSetGql from '@entities/set/api/graphql/mutation/CreateSetGql';
+
+export const CreateSetMock: MockedResponse<CreateSetResponse, CreateSetVariable> = {
+  request: {
+    query: CreateSetGql
+  },
+  result: (v) => {
+    const id = Math.max(...Object.keys(SetMockData).map(Number)) + 1
+    SetMockData[id] = {
+      ...v.sets,
+      id: id
+    }
+    return {
+      data: {
+        createSet: SetMockData[id]
+      }
+    }
+  }
+}
